backend/requirements.txt
# Ghost Creator SaaS — Backend Dependencies
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
pytrends>=4.9.2
feedparser>=6.0.10
requests>=2.31.0
google-genai>=0.5.0
pydub>=0.25.1
soundfile>=0.12.1
websocket-client>=1.7.0
Pillow>=10.3.0
playwright>=1.43.0
tqdm>=4.66.0
colorlog>=6.8.2
edge-tts>=6.1.9
elevenlabs>=1.0.0
google-cloud-texttospeech>=2.16.0
kokoro>=0.9.0
fal-client>=0.4.0
replicate>=0.25.0
numpy>=1.26.0
python-multipart>=0.0.9
supabase>=2.4.0
